Switch from using cloned escaper code to using Guava.
It's now possible to address this TODO because //third_party/java/android_libs/guava_jdk5 has been updated to a version which has PercentEscaper. b/11750512.

Note: The first attempt at this CL, at [] was rolled back. This new CL is the same, but the dependency which was broken is fixed in []

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=132233121