Improve iBooks buy e-book experience on iOS / Mac Chrome browser.
In [] I changed all the iBooks offer links to be custom URL scheme "itms-books://" from the pathquery. We have double checked with the Calypso team that it's good to show them.

In this CL, I added a URL scheme in the whitelists (both in the deeplinks util and SafeUrls ). In order to show the links on the KP, I added a repeated field allowed_custom_scheme in the LinkOptions and put [ITMS_BOOKS] in the third_party_action.html. In link_impl.html, I used the safe_type_service to sanitize URL with the allowed_custom_scheme.

A demo together with [] http://0.read-action-menglin-ibooks.fe.[].menglin.ij-q.[].google.com/search?q=to+kill+a+mockingbird&deb=0imobile. The iBooks offers are only shown on Mac or iOS.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=141505199