Added attributes to []html/types/codegen/html5_contract.textpb
The only manual changes are to that pb instance, and the rest were
generated by running

[]/> []html/types/codegen/update_generated_builders.sh

In [] xtof said

> <media src> should be safe?  I think we have elsewhere required TRUrl
> conservatively, but it's probably worth revisiting?  Or can <media> execute > flash (I hope not).

where <media> was a catchall for media elements in a table of elements.

This change is not required for that CL though since I don't think I can use html5_contracts.textpb as-is as the sole source to autogenerate the escaper tables because if every HTML5 attribute that allows interpolated values in polymer appears in this list as an explicit grant, then we will blow up the size of the HTML builders.
Tested:
Added tests to []google/common/html to test that
<video src> can be used with a SafeUrl and <iframe> still cannot.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150437650