html5_contract_textpb: Add attributes needed by polygerrit-ui
I integrated poly-resin in a local gerrit build and turned on
logging whenever a contract violation was detected.

There were only two on the CL list, and in the CL and diff view.
Both are boolean attributes.

---

We might want to define an Attribute.Contract value for boolean
attributes.  This CL does not do that.

For disabled, we could define it as ENUM with possible values ("disabled")
but the thing about boolean attributes is that the presence of the attribute
is the important thing, not the value.

Knowing this could allow API builders to leave out the value parameter for
boolean attributes.

---

Tested:
I read through the HTML5 spec to check that all values are equivalent,
hence the NONE is appropriate.

I reran tests for the regenerated builders and polymer_resin.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153372140