Reenable htmltypes/opensource tests on TAP
This reworks the script that unpacks the opensource tarball and runs
its tests to run on forge.

There are threads from 1-2 June 2017 on protobuf-discuss@ and build-help@
as to why the hack is the way it is, but I don't pretend to understand it
and hope that someone who understands ELF will point out the underlying
problem.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158133944